requirements.txt
emailer @ git+https://github.com/northriverboats/emailer.git
Envelopes
openpyxl
pyinstaller
pymssql
python-dotenv
workdays
